[Silver III] Title: △, Time: 12 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/27966. △/△.cc
@@ -7,7 +7,7 @@ int main() {
   long long n;
   cin >> n;
   cout << n - 1 + (n - 2) * (n - 1) << '\n';
-  for (int i = 2; i <= n; i++)
+  for (long long i = 2; i <= n; i++)
     cout << "1 " << i << '\n';
   return 0;
 }
